Philippine Standard Classification of Education (PSCED)

Philippine Standard Classification of Education (PSCED)

Usage

get_psced(
  ...,
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)

Arguments

...

See ?dplyr::filter. Expressions that return a logical value, and are defined in terms of the variables in returned data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.

token

Character. API access token.

version

Character. Version of the PSCED dataset. Default is NULL. If NULL, the latest version is used.

level

Character. Classification levels such as "all", "levels", "broadfield", "narrowfield", and "detailedfield". Default is "detailedfield".

harmonize

Logical. If TRUE, formats and standardizes the returned data. Default is TRUE.

minimal

Logical. If TRUE, returns a simplified dataset. Default is TRUE.

cols

Optional. Character vector of additional columns to include when minimal = FALSE.

Value

A data frame of PSCED classifications.

References

https://psa.gov.ph/classification/psced

Examples

if (FALSE) { # \dontrun{
get_psced(token = "your_api_token")
} # }
# If token is not provided, the function will fetch from local cache or
# download the latest version from remote repo
psced <- get_psced()

# Get specific level
psced_filtered <- get_psced(level = "levels")
psced_filtered
#> # A tibble: 9 × 2
#>   value label                                 
#>   <chr> <chr>                                 
#> 1 0     Early Childhood Education             
#> 2 1     Primary Education                     
#> 3 2     Lower Secondary Education             
#> 4 3     Upper Secondary Education             
#> 5 4     Post-secondary non-tertiary education 
#> 6 5     Short-cycle tertiary education        
#> 7 6     Bachelor level education or equivalent
#> 8 7     Master level education or equivalent  
#> 9 8     Doctoral level education or equivalent